:recycle: Split ActivityPub routes into directory module by endpoint responsibility
Pure mechanical move of server/src/route/activitypub.rs into a directory module; no behavior change.
